Refactor FluentValidation request processing
Updated `AddFluentValidationRequestPreProcessor` to use
`OpenRequestPreProcessor` instead of `OpenRequestPostProcessor`,
shifting to pre-processing of requests.

Modified `FluentValidationRequestPreProcessor` to accept
only `TRequest` as a generic type parameter, removing the
need for a response type and simplifying the interface.
The constraint on `TRequest` was also updated to implement
only `IRequest`, focusing on request processing.