Validate client and unprotect user credentials
Added logic to retrieve and validate the `client` associated with
the `user` using `Context.Clients.FindAsync`. Introduced a null
check for the `client` to return `UnitResult.NotFound<UserDto>`
if the `client` is not found. Integrated a call to
`userCredentialProtectionProvider.Unprotect` to unprotect the
`user` with the validated `client`. These changes ensure proper
validation and secure handling of user credentials.