Add SuspendedTimestampUtc to AccessToken table
Introduce a new nullable column `SuspendedTimestampUtc` of type `DateTimeOffset` to the `AccessToken` table via a new migration class `Added`. This change includes methods to apply and revert the migration. The `SuspendedTimestampUtc` property is also added to the `DbAccessToken` entity model in the `CFIdentityDbContextModelSnapshot.cs` file to reflect this new column in the entity framework model.