Add Authorise class for role-based authorization
Introduces a new static class `Authorise` in the
`IDFCR.Http.Authentication.Extensions` namespace.
The class includes a generic method `Using<T>` that
creates an `IAuthorizeData` object based on a
specified `RoleCategory` and additional roles.
This method ensures type safety by requiring `T`
to implement the `IRoleRegistrar` interface and
calls `RoleRegistrar.FlattenedRoles<T>` to
retrieve the appropriate roles for the
`AuthorizeAttribute`.