Add lastAddedEntry tracking in RepositoryBase
This update introduces a private field `lastAddedEntry` to track the last added entity entry in `RepositoryBase.cs`. The field is of type `EntityEntry<TDb>?` and is set to `null` after saving changes in `SaveChangesAsync`. The `UpsertAsync` method now assigns the result of `DbSet.AddAsync(dbValue, cancellationToken)` to `lastAddedEntry` when adding a new entity. A new protected property `LastAddedEntry` has been added for access to this field. Additionally, a call to `BeforeUpdate(foundEntry, value)` has been included before updating existing entities.